WORK/8.1_self_understanding/entertainment/requirements.txt
gigachat>=0.1.0
pymorphy3>=2.0
requests>=2.28
python-dotenv>=1.0
